Ensure consistent use of pipe U-values.
In the input file defining the U-values of the available pipes, the user can
provide the columns 'f_length_loss' and 'f_length_cost'. If they do not exist,
a default value of '1' is assumed for each DN. This means the U-value already
describes both forward and return flow. This applies e.g. to manufacturer data
of double pipes. For single pipes, a f_length_loss value of 2 can be given,
if the U-value of a single pipe is given. The U-value will then be multiplied
by 2, internally.
If costs desribe the costs per meter trench length, f_length_cost 1 is the
correct choice.